Colour Randomizer: New randomization method
New method randomizes entire body and for every vehicle as well,
including taxis and cop cars, which were previously not
randomized. This also makes sure that the new colour indices of the
vehicles are not saved to the save file. Furthermore, the mission
vehicles will no longer have a consistent colour on consecutive
mission starts as was previously the case where the randomization was
only seen once, and the colour remained the same for the rest of the
game session. It will also now randomize from a truly random set of
colours, making all combinations of colours possible. The Pay 'n Spray is no
longer restricted to the colour table defined by the original game.

This build also fixes the Mission Randomizer bug with Bleed Out, where
completing Bleed Out (marker) would not clear its blip until the
actual Bleed Out (randomized) is complete.